Update pair initializations, reorganize GOB/Event code
Optimized builds of the previous commit uncovered a bug, which came
from static analysis based on the trashing done by RESET_CELL() vs.
merely RESET_VAL_HEADER() and leaving the contents ambiguous.  That
bug was that a pairing series was being used under the assumption of
both pair components residing in a single cell.

(Ren-C eliminated the idea of a PAIR! that fit into one cell, as it
introduced new numeric categories besides INTEGER! and DECIMAL! to
deal with...since both of those require more than half the cell to
represent.  Instead, it established a form of REBSER node that could
act as two values--thus being a commonly allocated size and also only
requiring one pooled allocation vs. a node and then series data.)

While fixing that, this goes ahead and makes the SET_PAIR() construct
obey the Init_XXX conventions, as Init_Pair()...and updates code using
it (mostly GOB! and the GUI portion of EVENT!, which does not have
much testing at the moment, and are likely to be discarded or rewritten
completely--but serve as interesting test cases).