0 to the power of 0 is 1
While 0 to the 0 power is undefined in a mathematical sense, most
programming languages choose to make it 1:

https://rosettacode.org/wiki/Zero_to_the_zero_power

Beyond being a recommendation of Don Knuth, it's also in the C99 formal
standards for what `pow()` does, as well as the answer in Matlab.

So this replaces the previous answer of 0.0 with 1.0.