Rename "Type Actions" to be "Generics"
When the various ANY-FUNCTION! types were unified, they had to take a
single name.  While FUNCTION! might seem a natural choice, the fact
that FUNCTION is a generator meant this pattern caused one to lose
access to the generator:

    block: [...]
    integer: 10
    function: :add ;-- oops, can't do f: function [...] [...]

So this meant picking ACTION! from among the previous kinds of function
to be the "one name" made sense.  It has other benefits (including
being less to type).

However, internal to the code there was still a need for a name for
the category of implementations that ACTION! had related to.  Since
they were dispatching common code to per-datatype code, the name
"Type Action" was chosen at first.  However, the Factor language and
Lisps would call these "Generics":

https://en.wikipedia.org/wiki/Generic_function

It seems like a good name change, so this goes ahead and does it...
though for now, there's no real impact on userspace.

This also goes ahead and makes the conventions of the natives "details"
array organization clearer.