Reimagine UNLESS as OR/XOR variant, GROUP!s in LOGIC ops
An experiment to try and rethink UNLESS had made it into an infix
operator which would evaluate to its left hand side expression unless
the right hand expression was non-NULL.  This was intended as a kind
of "reordered default" so you could put the default first, and then
have the less common cases tacked along after that.

Due to the fact that the left hand side was always evaluated, however,
this conception turned out to be of limited use.  It was also a bit
confusing that the right hand side was NULL-active, when the word
"unless" sounds as if it is conditional on things being truthy/falsey.

This commit varies the idea a bit to make UNLESS act more like the
constructs in the family AND/OR/XOR:

    >> (print "a" <left>) unless (print "b" null)
    a
    b
    == <left> // same as OR

    >> (print "a" null) unless (print "b" <right>)
    a
    b
    == <right> // same as OR

    >> (print "a" <left>) unless (print "b" <right>)
    a
    b
    == <right> // DIFFERENT! right result wins out

This also goes ahead with the change that makes GROUP! on the right
hand side of AND and OR subvert the short-circuiting, which is now
only done when there are BLOCK!s.  XOR and UNLESS then only allow
GROUP! on the right, to underscore the fact that they don't make
their decision short-circuited.