First phase of moving console to an extension
This takes the Read-Eval-Print-Loop and hacks it around to make it a
CONSOLE function.  The intent is to make it possible for the REPL to
be reused in both the main() function, as well as from inside the
BREAKPOINT naive, that is part of the debugger extension.

There are several complicating factors for this idea, and this only
does the minimum to make it possible to run nested CONSOLEs.