Reverse sense of new-line markers during REVERSE
Newline markers are on value cells in arrays, and signal that a newline
should be output *before* that element during rendering.  (Ren-C has
an additional attribute on arrays indicating if a newline should be
rendered before printing the closing parenthesis/bracket.)

Historically, the REVERSE command left these markers untouched during
a reversal.  e.g.:

    >> data: collect [keep/line [1 2] keep/line [3 4 5] keep/line [6]]
    == [
        1 2
        3 4 5
        6
    ]

    >> reverse data
    == [
        6 5 4
        3 2
        1
    ]

The results were less useful than a method which would reverse the
sense of the newline markers...e.g. those elements that had previously
been at the start of a line would now be "reversed" such that they
would be the end of a line:

    >> reverse data
    == [
        6
        5 4 3
        2 1
    ]