Remove debugger hook to call CONSOLE via usermode
Debugging functionality needs to spawn a nested REPL.  One wouldn't
want to rewrite the entire console functionality (shortcuts, skinning,
multiline input, etc.) so it needed to reuse the code.

Previously this was done with some C-level pointer-to-function hooks.
With the console moved to an extension, this makes the hookup via
calling a Rebol function via rebRun() instead.

Some experimental debug features are removed in order to reduce the
code complexity, to better approach a "minimum viable product" as far
as breakpoints and backtraces are concerned.