Remove integer mapping for built-in errors, use WORD!s
While help pages for built-in errors dispatched based on the error's
word ID, R3-Alpha also gave errors an integer code number.  These
numbers were understood to be "volatile", as adding and removing errors
during the development process would juggle their values--and it would
leave unnecessary gaps and mess to try and maintain them.  But the
theory seemed to be that at some point, the system would "harden" and
the numbers would be fixed, so that protocol design could use the
integer values in lieu of the names.

In practice, there is very little compelling reason to involve Rebol
in the integer-level standardization of its error codes.  The system's
fundamental unit of exchange is Rebol-structured data, and building
error IDs from WORD!s, PATH!s, or URL!s makes much more sense than
trying to codify numbers based on it.

One useful area for having numbers--however--was for the invocation of
Rebol-structured errors from C code.  These RE_XXX constants ("Rebol
Error") were parameters to the variadic error constructor.  This commit
rethinks that to use SYM_XXX values of the error IDs and categories
instead--the same numbering mechanic used for any special word known
to the system (e.g. %words.r).

A fledgling facility for adding new errors was put into the extension
mechanism.  Which is a useful thought--however, it was built on the
RE_XXX numbering system.  Instead, it should use a facility that would
register errors to the catalog that would be available to any script
user, and then raise those errors with libRebol API calls that would
also trigger those errors as users would (vs. expose the internal API
fail() mechanic).  This switches those calls to use rebJumps() with
a call to the usermode `fail` function, giving the same legibility
without a guaranteed ID scheme (which the RE_XXX mechanism did not
have either, as it depended on module load ordering).

This allowed significant simplifications, paving the way for systemic
improvements to the error machinery.