Step in simplifying/standardizing extensions
There were two key technical differences in the extension model that
Ren-C aimed to have that R3-Alpha's did not.  One was the ability to
choose to build extensions into the executable instead of as DLLs to
be loaded after the fact if one wanted.  The other was that writing
natives was done using specifications in comments, with the same API
that natives internal to the system used (instead of having a separate
COMMAND! interface).

These two basic goals were accomplished, however there were many ideas
that got carried in along the way.  Since originally there was very
little Rebol code in extensions (e.g. GIF/JPEG encoders/decoders or
cryptography) there was a limited facility to include an uncompressed
string of UTF-8 Rebol code as a literal in the extension's C file.
This remained even after the need arose to have a full Rebol file
compressed and embedded.  There was also the idea of multiple modules
per extension, which creates a fair amount of complexity.

This takes a simplifying step to start standardizing the modules on a
manageable common denominator.  All extensions have a Rebol file which
is compressed in.