Temporarily move to one-module-per-extension
The original concept of extensions was to support multiple modules.
That's still likely a good idea, however since users are not going
to interact with an EXTENSION! data type, the emphasis on extensions
had clouded matters where the module system needed to be providing
features that applied to either usermode modules or native modules
(when the truth is many would be hybrids).

Natives need to know which module to use as the binding context for
material loaded from strings in API calls.  This is important for
many reasons--such as if a module is written to Rebol2 conventions
instead of Ren-C, then `rebRun("append", block, "item");` has
different behavior.  However, being able to poke the right module
context into the natives was extremely difficult to do as written.

Ultimately, the feature is still interesting--just as the EXE
itself can ship with multiple embedded modules, an extension should
be able to as well.  But to move the focus to modules, this
temporarily makes it so each extension contains one module...just
to help move the design forward.