Remove ASSERT-NO-BLANK-INSIDE from rebmake
It used to be pretty easy to accidentally get BLANK!s (NONE!s) inside
of blocks.  You could get them with `compose [a (if false ['b])]`, or
have a BLANK! showing up unwanted almost anywhere as the outcome of an
ANY or ALL or SWITCH or SELECT that didn't work out how you thought.

Seemingly due to frustration, a routine called ASSERT-NO-BLANK-INSIDE
was added to rebmake as a sanity check run on blocks that were part of
the make configuration.  The test helped avoid blanks that were
inadvertent products of evaluation getting treated as purposeful values
by code that might react to it in places like conditionals.

But in modern Ren-C this is basically non-existent.  NULL can't be put
in blocks, and you don't accidentally make BLANK!s; you have to
explicitly do something like a TRY to get them.  Such a routine is
no longer necessary.

Includes a patch for a test that was failing on Linux due to a change
to the bracket-form of AND, as well as renaming %make-ext-natives.r
file to the more meaningful %prep-extension.