Streamline rebmake "class-name" to class w/ISSUE!s
The %rebmake.r file introduces prototype OBJECT!s for representing
various build artifacts (obj files, applications, build variables, etc)

(It seems to be roughly based on CMake's model, using some of its
terminology--"object libraries" vs. "static libraries" for instance.
This likely comes from the fact that the original code Shixin was
working on was generating CMake files.)

It was a bit verbose, because the field was called "class-name" and
it used a WORD! like "object-library-class", resulting in things like:

    if foo/class-name = 'object-library-class [ ... ]

This just does some basic shortening to make it a little more readable
and ergonomic, such as:

    if foo/class = #object-library [ ... ]