Add COUNT-UP, COUNT-DOWN, and experimental SO
   >> count-up n 3 [print [n]]
   1
   2
   3

   >> count-down n 3 [print [n]]
   3
   2
   1

For now, COUNT-UP and COUNT-DOWN are specializations of FOR...just to
keep the total amount of code in the system down.

Long term, the COUNT-UP may be part of a plan to repurpose REPEAT,
which has an additional function besides counting up if the argument
is a series--so that feature needs to be addressed somehow.

Includes an experimental postfix-assert application of SO:

    >> 1 = 1 so print "all's well"
    all's well

    >> 1 = 2 so print "won't happen"
    ** Error: ...