Update MSVC project to current WinSDK shipping w/VS2017
If you build a MSVC project file with a different Windows SDK version
number than your installation has, it will ask if you want to retarget
the solution as the first thing it does when it loads.  This isn't
a huge deal but is somewhat annoying.  This commit bumps it up to
the current version you'd get installing MSVC 2017 off the web today,
so it will ask those with older versions to retarget--not the current.

Since this was done while debugging an unrelated problem, there are
some changes which help make the generated output from the build
process match up better with the exact spacing and casing that MSVC
will spit out if it writes the XML out again (e.g. what it does when
it retargets SDKs...)

Contains some other various simplifications while browsing the code,
such as taking advantage of the already-established but becoming more
guaranteed ability of APPEND to act as a no-op when NULL is used as
the second argument.