Reimagine extension DLL hook as "collation"
This is a first step toward simplifying the extension mechanism to
push more of the process of native registration into the core, while
moving the binding and export process to reuse usermode module code.

One part of that is to reduce the number of pointers being passed
around down to just one API REBVAL* returned per extension...a
so-called "collation".  At the moment this collation is a BLOCK! of
HANDLE! values that can be used by the system (in INIT-EXTENSION) to
produce a live MODULE!.

A casualty of this simplification--for the moment--is that the unload
process hasn't been formally defined.  At minimum it needs to be,
because some modules might initialize resources with malloc() or
otherwise, that need to be freed before they are shut down.  However,
this mechanism could apply just as well to user-written modules that
are unloaded (they may need to close network connections or do other
business for normal shutdown).  So this feature is punted on in order
to design a mechanic that could be applied to any module.