Allow null to unset a SET-WORD! or SET-PATH!
This restores the behavior from the initial development of null, which
allows it to unset variables:

    >> x: 10
    == 10

    >> x: null
    // null

    >> x
    ** Script Error: x has no value

However, VOID! values (which unlike null *are* values and may appear in
arrays and can be accessed without a GET-WORD!) still cause errors:

    >> x: do []
    ** Script Error: x: can't be VOID! (use TRY, OPT, or SET*)

    >> x: print "hi"
    hi
    ** Script Error: x: can't be VOID! (use TRY, OPT, or SET*)

Allowing null assignments to mean "unset" provides convenient means of
removing keys from maps, which could be difficult otherwise:

    >> m: make map! [1 a 2 b]
    == make map! [
        1 a
        2 b
    ]

    >> m/1: _

    >> m
    == make map! [
        1 _
        2 b
    ]

    >> m/1: null

    >> m
    == make map! [
        2 b
    ]

The original intent of trying to raise errors on null was to improve
error locality, drawing attention to the moment where an oversight
in handling null conditions happened.  But analyzing the pros and cons
suggest that the only real outcome of this process were kneejerk
mitigations which probably create more opportunities for bugs than
they solve.  Additionally, it led to constructs which would seem clean
and easy to write in Rebol2 seeming to need extra words thrown in for
little benefit.

It is presumed that having the variables assigned be unset--and thus
difficult to work with if used after the point of assignment--is enough
"safety", balanced against the convenience of being able to perform
such assignments.  Also, the flexibility of usermode constructs to be
adapted to catch patterns of common errors in any given codebase which
finds itself plagued by problems related to null performing unsets.