Drop ZeroMQ dependency on old x86 linux installation
The cross compiler was installing a ZeroMQ dependency which stopped
being supplied by the apt repository sometime in Nov 2018 and breaking
the build.  As it happens, the configuration doing this build didn't
even use ZeroMQ because it is C89, and the ZeroMQ extension was
updated to use the C99 features of libRebol...so it was just breaking
the build for no good reason.