DELIMIT (and UNSPACED, SPACED, etc.) null if all opt out
This is a change which was contemplated for some time (as early as
the proposal for "COMBINE") which is to distinguish the situation of
a block-to-string conversion that entirely opts out by returning a
null to indicate opting out.

    >> unspaced ["a" if false ["b"] "c"]
    == "ac"

    >> unspaced [if false ["a"] if false ["b"] if false ["c"]]
    // null

Previously an empty string would be returned.  This can still be done
a number of ways, one of which is to provide any empty string as
input in the block:

    >> unspaced ["" if false ["a"] if false ["b"] if false ["c"]]
    == ""

But it can also be done with ELSE, UNLESS, or any other null-reactive
construct.  By doing it this way, empty string isn't the only option,
so for instance if you want to comply with the C standard for function
arguments you aren't supposed to have an empty paramlist as just
`int foo() {...}`, but `int foo(void) {...}`

    >> params-text: "void" unless (delimit c-function-params ",")

Given the options--in particular the option of just putting `""` in
the input block being enough--the advantages of being able to uniquely
react to the empty state are high enough to be worth it.