REPLACE allows null to opt out of pattern or replacement
This modifies the behavior of the CHANGE function w.r.t. NULL when a
/PART refinement is used.

Previously, it would consider a NULL to signal a no-op...now it
considers it the absence of a replacement.  This makes it similar to
a /PART when splicing an empty array:

    >> head change/part copy [a b c d] [] 2
    == [c d]

    >> head change/part copy [a b c d] null 2
    == [c d]

This is applied for BINARY! and STRING! as well.  Since REPLACE is
a mezzanine based on CHANGE, this means it now can support a null
pattern (signaling a no-op) or a null replacement (meaning to just
vaporize any instances of the pattern)

    [3 4] = replace copy [3 0 4] 0 null
    [3 0 4] = replace copy [3 0 4] null 1020
    [2 0 2 0] = replace/all copy [1 0 2 0 1 0 2 0] [1 0] null
    "34" = replace copy "304" "0" null
    "304" = rplace copy "304" null "1020"
    "2020" = replace/all copy "10201020" "10" null
    #{3040} = replace copy #{300040} #{00} null
    #{300040} = replace copy #{300040} null #{10002000}
    #{20002000} = replace/all copy #{1000200010002000} #{1000}