Simplify LIST-ENV and fix windows CMD.EXE shell environment
There is a quirk of launching Windows executables from a shell which
puts odd environment variables into the returned list.  See the article
"What are these strange =C: environment variables?"

https://blogs.msdn.microsoft.com/oldnewthing/20100506-00/?p=14133

These were not handled in the original Windows version of LIST-ENV
added to open source R3.

In addition to solving this issue, this converts LIST-ENV to use
libRebol APIs.  While this may be a little bit less efficient, the
performance of getting the full list of environment variables is not
a core concern...and in fact, the proper interface for the environment
is probably as a virtualized MAP! object, so you could say:

    for-each [key val] environment [
        ...
    ]

In any case, it's more important to get good tests of the libRebol API
and remove unneeded routines, vs. have a slightly more efficient
LIST-ENV at this particular time.