Small-scale test of SER/ARR terminology change
Without modifying the entire codebase just yet, this tries on the idea
of using "Ser" and "Arr" in function names when it operates on the
raw series and array nodes vs. values.  It just changes Make_Ser()
and Make_Arr() and the relevant return results.

If this turns out to feel like a good change, then it will be pushed
through further in the system.

https://forum.rebol.info/t/905