Attempted improvement at invisible function compositions
The invisible mechanic behind COMMENT and DUMP hasn't has several known
holes in the evaluator model which will trigger asserts.  This goes
for questions like how a single-step EVAL should handle an invisible,
or how to call invisibles with refinements via SHOVE.

But some problems are fairly easy to address, such as getting ADAPT
or SPECIALIZE of an invisible to work.  This improves some of those
mechanics by factoring a bit of code in %c-eval.c, though a general
review is needed to see what it would take to clean it up and get it
more fully "right".