-- as DUMP, ?? as PROBE, various DUMP enhancements
This renames the shorthand for DUMP to `--`, making it more convenient
than ever to expect variables basically anywhere in the code (as it is
invisible to the evaluator).  Not having to use the shift key to type
the dashes is helpful, and it also allows REFINEMENT! arguments for
easy output.

    >> all [-- /a 1 2 3 false -- /b] print "^ didn't make it to /b"
    -- /a
    ^ didn't make it to /b

Also makes it more possible to SPECIALIZE and ADAPT the DUMP operation
without needing to pass variadic input, by quoting the first argument
and achieving its special features using a second variadic argument
which can be left empty.