Make CONTINUE "endable"
Ren-C initially introduced the idea of a /WITH parameter to CONTINUE,
so that a CONTINUE operation would be able to make it seem as if that
iteration of the body had returned a value.

However, since CONTINUE is at the tail of an operation, it's possible
to take advantage of that by having it be able to act as arity-0 or
arity-1 depending on the presence of an argument:

    >> map-each x [1 2000 3000] [
           if x > 1000 [continue]
           x + 1
       ]
    == [2]

    >> map-each x [1 2000 3000] [
           if x > 1000 [continue <big>]
           x + 1
       ]
    == [2 <big> <big>]

This parallels the behavior of RETURN, which allowed to repurpose the
EXIT command.