Fix bug where enfix considered invisible branches stale
When you call Eval_Core(), you can "prime it" with a value in the out
slot that will be the return result if nothing evaluates.  So if you
wanted, you could prime the output cell of a frame with 10... run
an evaluation of `[comment "hi"]` (or just `[]`), and when the eval
was finished it would still hold 10.

But the primed value is not a candidate as an enfix left parameter.  So
if you preload with 10, and then evaluate `[+ 20]`, it will be an
error just as if you hadn't primed it and ran `[+ 20]`.  The only time
you should have the 10 coming into play is if there was really nothing
(e.g. all invisible or empty) that evaluated.

The way the system tracks whether the value in the output cell is
a legitimate enfix candidate or not is the "stale" bit.  This bit is
available to callers of the evaluator...e.g. they can prime it with
10 and then run `[5 + 5]`, and test the stale bit of the result to
tell if it was the 10 they passed in, or if it had been overwritten
by a new evaluation.

The stale bit is not copied around when values are moved...it is masked
out.  But when various levels of the evaluator run, they often all
reuse the same output location if they can.  This means one level's
stale bit could wind up confusing another level's if it's not cleared
off when deemed "uninteresting".

This was the case with `if true [comment []] else [...]`.  The branch
would be primed with a void by IF, be mared stale by the evaluator, yet
the same frame would then run on the same output value and try to do
the ELSE.  Having the stale bit set would mean the ELSE would reject
the void on the left as not a candidate, and act as if it were just
`else [...]`, missing a left hand value.

The solution is that at the higher level granularity of running arrays,
the eval result needs to clear the stale bit off.  Also adds a check
to the evaluator to notice when a function dispatcher says it's
returning a value (vs. being invisible or redoing) and making sure
that such values don't have the stale bit--a check that would have
caught this kind of problem in the first place.

This bug was actually observed during the development of DUMPS, but
unfortunately worked around and forgotten about...hence the workaround
was committed.  :-/  This also removes that workaround, and throws in
a new DUMPS feature (bug was rediscovered while adding it) that lets
it implicitly create a dumper for a SET-WORD! at point of declaration.