Change protocol for most loops to BLANK! for never running
The initial definition of the loop result protocol used BLANK! for a
loop that never ran, and NULL for a loop that would break...then
forced all loops that ran at least one iteration of the body and did
not break to a "truthy" value.  This gave rise to the controversial
"barification".

This changed to having the loop primed with VOID! and moving the
responsibility for truthy/falsey responses to the loop body.  But
it turns out to be much more useful to adopt the idea that a loop
never runs as BLANK!, and then voidify any loop bodies that try to
return *either* blank or void.  This puts the user in control, and
gives an easily comprehensible error if their loop body returns a
blank or a void and they try to operate on it conditionally.