Centralize stale bit handling
The "invisibles" mechanism introduced a lot of handling in the
evaluator, which relied upon asserts to signal when things were going
wrong.  It was known that there were a lot of "dark corners" which
would be addressed as they became important by appearing in usage.

These usages are now on the rise, with invisibles playing a larger
and larger role (e.g. with DUMP and specializations/adaptations of it).
This commit builds on the earlier commit that adds a centralized
assert finding "leaky" stale values, to add a DO_FLAG which is needed
to instruct the evaluator not to clear off the stale bit...and
quarantines those cases to their own special primitive.

It's somewhat clunky at the moment, but hopefully sets up a perimeter
such that a more efficient solution can be found later.