Eliminate many superfluous TRYs
It was decided that `x: null` should unset x instead of causing an
error.  This means that significantly fewer places in the code need
a TRY, if they are of a form like:

    if x: try (...) [
        // only code that uses X
    ]
    // code that is not supposed to use X

In fact, it is *safer* to unset the variable than to set it to blank
for this kind of scenario, because it captures bad usages after the
branch.

Because much of this code is in the bootstrap process, any R3-MAKE that
is since the SET-WORD!-of-null-causes-errors change and the change that
they no longer do cannot build this.  Hence an R3-MAKE which is old
(such as the ones committed in the make directory) or one that is
relatively recent will be needed.

This commit includes some small cleanups, e.g. to TO-C-NAME