VARARGS! new-line detection ** comments to end of line
This adds the ability to tell if a varargs is at a newline boundary,
and also introduces a construct to use it--the `**` comment.  The `--`
is also changed to dump to the end of line as well.

It can also be stopped by reaching the end of a GROUP! or BLOCK!, or
hitting a BAR!.