Reclaim THROWN() bit in value header
R3-Alpha indicated thrown-ness on a value by setting a bit on it.  This
was tested with the THROWN() operation.

One of the dangerous aspects of this is that thrown values might be
processed normally, with their thrown-ness ignored.  Ren-C added audits
for this by naming routines that could potentially throw with the
_Throws() suffix and having them return a boolean, to hint callers that
they needed to write `if (Some_Operation_Throws(...)) {...}` and
react to the condition.

This led to a desire to get rid of the thrown bit, as it doesn't make
much sense to use a value bit for something that can only be on one
cell at a time in the evaluator--it reflects more a state of a frame.
But getting rid of the thrown bit would carry some danger as places
couldn't check to make sure the thrown-ness was being ignored when
it shouldn't be.

So this commit pushes the return value piping of throw status more
deeply into the system, fixing some hacks (like how MAKE BLOCK! from
a VARARGS! had to be handled from MAKE itself in order to throw).  It
also more pointedly shows dangerous cases, like calling MAKE as
construction syntax inside the loader--which should not be running
any usermode code.

For this phase, checking the thrown bit is only an assert--and is not
necessary for functionality.  Next step will be to reclaim the bit.