COLLECT w/no KEEPs (that don't KEEP a NULL) is NULL
This switches it so that:

    >> collect []
    // null

It saves on the creation of blocks, and also permits TRY COLLECT to
give blank, which for enumeration purposes serves as well as an empty
block.  If a block is desired regardless of keeps, an idiom for doing
that is:

    >> collect [keep [] ...]

The result will always be a block, because even a spliced keep of
nothing is considered keeping "something".  However, KEEP of a NULL
does not count:

    >> collect [keep if false [<not kept>]]
    // null