Fix MSVC project generation, plain FAIL in switch []
While MSVC lacked a STRIP command, it was actually returning something
and running it, so the use of a `...` marker for unimplemented would
stop the r3-make process from running.  This adds a debug message and
that should be looked at later as to what the intended protocol was
for not having a command (here it returns an empty string).

Also uses the idiom of having SWITCH statements go directly to a FAIL
instead of calling `default [fail ...]`, which makes the code slightly
more compact.