Report expired frame assignments in SET-WORD! assignments
The idea of expiring series is that their data allocation can go away,
leaving only the series node (sizeof 2xREBVAL) to track it and prevent
dereferences from crashing.  The point at which testing for an expired
context was being done was at VAL_CONTEXT(), which extracted a context
from a REBVAL*.

However, another entry point for getting contexts from REBVAL is to
look them up from words--which might point to stale references.  This
code path was not being checked, which led a typo that intended to
point to a variable that outlived a function to point that one that
didn't.  Assigning that variable crashed, instead of giving an error.

This fixes the variable assignment to use the right variable and also
changes the behavior to give an error instead of crash.

Also fixes variadic DUMP of GROUP!s via `--`, found while debugging.