Make varlists for FRAME! pass ASSERT_SERIES
While debugging it was encountered that fixed-sized arrays (such as
those produced by Push_Action and the evaluator for frame varlists)
don't always have valid bits in their "rest" capacity.  This isn't a
problem--those bits are supposed to be uninitialized.  But dynamic
arrays do need those bits initialized in case of growth.  The
difference wasn't being taken into account in the debug routines that
check arrays for validity.

This commit fleshes out the debug invariants.  It also updates the
termination of arrays so it doesn't require a conditional test of
whether to write a zero byte to the end signal slot, due to changes
in the rules of the system where cells must be formatted to write.