Make INPUT return null on cancel, DUMP fixes
This changes the INPUT function to return a null when it is canceled
via Escape (as opposed to being Ctrl-C'd, in which case the script
stops and it returns nothing at all).  Due to technical limitations,
Windows input cancellation is done with Ctrl-D, and this is then
supported on Linux as an alternative.

Also includes some fixes to the DUMP routine in dealing with PATH!s.