Eliminate stray "+ 1" in array copying
Some of the array copying code had put in a "+ 1", presumably for an
end marker at one time.  The end marker is already accounted for in the
logic of array creation.  This was just considered excess capacity and
not a bug, until a recent commit that checked for stray cells in fixed
size arrays noticed the mistake.