Unify "Facade" with Paramlist
When specializations were first created, it was clear that the newly
created function would need a parameter list with fewer parameters than
the original.  However, when the function ultimately ran it would need
to have a frame with the full number of parameters as the original
function in order to execute the code.

The initial concept was just to store the "Underlying" function and
create a paramlist with fewer arguments.  Then switch over to the
underlying function for frame fulfillment.  However, this did not
allow for the idea of tweaking parameter types or argument conventions
(e.g. quoted or non, tight or not) in a derived function.  So the
idea of a "facade" was introduced.

The concept was to create facades only when needed, but in terms of
dealing with system completeness it started to seem every action would
need a facade.  But what makes more sense is to not shorten a paramlist
for a specialization, but just to *hide* parameters in it.  This would
let each derived action's paramlist serve the purpose that a facade
did, and simplifies needed explanations significantly.

This also improves performance, as switching over from paramlists to
facades ate a lot of CPU time.