Rethink FOREVER construct as CYCLE, add STOP ability
The FOREVER construct seemed on the surface as a more efficient and
more fluent way to write `while [true]`.  But its name is long-ish, and
actually rather jarring for casual use of "loop until some condition
sensed within the loop causes a break".  That is not sensibly called
"forever"--so it didn't really seem to get used much at all.

CYCLE is a shorter name, without that baggage.  "Keep cycling until
something stops the cycle".

It also adds a distinct feature of STOPping cycles with a result value.
While loops aren't allowed to "return a value from a break", that's
because they need to distinguish a stopping request from normal
completion.  However, *CYCLE has no normal loop termination condition*.
It isn't counting anything, or testing anything.  So its return result
would go somewhat to waste if there wasn't a way to give one.

Rather than tamper with the BREAK-is-null concept, this adds a unique
variant called STOP that only CYCLE will respond to:

    >> x: 1
    >> cycle [
           if x = 3 [stop <stopped>]
           print [x]
           x: x + 1
       ]
    1
    2
    == <stopped>

STOP will not accept NULL.  So from the outside one can still tell a
BREAK condition from a "normal" stop of a cycle.

FOREVER is temporarily aliased to CYCLE; there may be a point to
keeping it around, though it should probably not be offering a BREAK
or a STOP if it's "forever" (!)