Add N-SHOT, ONESHOT, UPSHOT
This adds a function generator which is able to take an argument of
code to run, but it will only run that code N times.  If the N is
negative, it inverts it so it will only *not* run that code N times.

    >> once: n-shot 1
    >> once [5 + 5]
    == 10
    >> once [5 + 5]
    // null
    >> once [5 + 5]
    // null

    >> anti: n-shot -1
    >> anti [5 + 5]
    // null
    >> anti [5 + 5]
    == 10
    >> anti [5 + 5]
    == 10

These specializations are called ONESHOT and UPSHOT respectively.