Fix lost 'thrown' bit when starting new expressions
The START_NEW_EXPRESSION_MAY_THROW macro takes a goto for the place to
go on a throw (e.g. a Ctrl-C generating a HALT).  It is invoked in
two places, but only one of them was adjusted to jump to the
`return_thrown` label instead of `finished`.  This updates the other
instance.