Change TRAP to return NULL on no failure
Historically, Rebol's error-trapping function took a handler--which
could be a block or a single-arity function--for handling the case
when an error was raised.  If no handler was provided, it would
return the error value...but this would be mixed up with the possible
return value of the expression evaluated itself.

This rethinks TRAP to return NULL if there is no failure.  That means
it doesn't need to interact with handlers, you can use THEN and ELSE.
This does make it a little harder to get a return value from the
expression, but many expressions written as:

    foo: trap [...]

Can be rewritten as:

    trap [... foo: ...]